requirements.txt
aiohttp==3.8.4
aiolimiter==1.0.0
aiosignal==1.3.1
annotated-types==0.6.0
anyio==3.6.2
APScheduler==3.10.1
async-timeout==4.0.2
attrs==23.1.0
autopep8==2.0.2
black==24.3.0
cachetools==5.3.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.7
cryptography==40.0.1
distro==1.8.0
flake8==7.0.0
frozenlist==1.3.3
google-ai-generativelanguage==0.4.0
google-api-core==2.17.1
google-auth==2.28.1
google-generativeai==0.4.0
googleapis-common-protos==1.62.0
grpcio==1.62.0
grpcio-status==1.62.0
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==0.16.3
httpx==0.23.3
hyperframe==6.0.1
idna==3.4
mccabe==0.7.0
multidict==6.0.4
mypy-extensions==1.0.0
openai==1.1.1
packaging==24.0
pathspec==0.12.1
platformdirs==4.2.0
proto-plus==1.23.0
protobuf==4.25.3
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycodestyle==2.11.1
pycparser==2.21
pydantic==2.4.2
pydantic_core==2.10.1
pyflakes==3.2.0
pyTelegramBotAPI==4.10.0
python-telegram-bot==20.2
pytz==2023.3
pytz-deprecation-shim==0.1.0.post0
requests==2.28.2
rfc3986==1.5.0
rsa==4.9
six==1.16.0
sniffio==1.3.0
socksio==1.0.0
telebot==0.0.5
tomli==2.0.1
tornado==6.2
tqdm==4.65.0
typing_extensions==4.8.0
tzdata==2023.3
tzlocal==4.3
urllib3==1.26.15
yarl==1.9.2
